Crud Respositories Created

+++ src/main/java/com/example/whiteboardsp19likithponnanna/repositories/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 
 public interface UserRepository
-        extends CrudRepository<User, Integer> {
+        extends CrudRepository<User, Long> {
     @Query("SELECT user from User user WHERE user.username=:username AND user.password=:password")
     public List<User> findUserByUsername (@Param("username") String username, @Param("password") String password);
